Transform HTTP into a behaviour to be able to mock in Playground
